+++ packages/color-helpers/src/conversions/d65-to-d50.ts
@@ -1,5 +1,11 @@
 import type { Color } from '../types/color';
-import { multiplyMatrices } from '../calculations/multiply-matrices';
+import { Matrix, multiplyMatrices } from '../calculations/multiply-matrices';
+
+const M: Matrix = [
+	1.0479297925449969, 0.022946870601609652, -0.05019226628920524,
+	0.02962780877005599, 0.9904344267538799, -0.017073799063418826,
+	-0.009243040646204504, 0.015055191490298152, 0.7518742814281371,
+];
 
 /**
  * Bradford chromatic adaptation from D65 to D50
@@ -12,11 +18,5 @@ export function D65_to_D50(XYZ: Color): Color {
 	// - convert from XYZ to retinal cone domain
 	// - scale components from one reference white to another
 	// - convert back to XYZ
-	const M = [
-		[1.0479297925449969, 0.022946870601609652, -0.05019226628920524],
-		[0.02962780877005599, 0.9904344267538799, -0.017073799063418826],
-		[-0.009243040646204504, 0.015055191490298152, 0.7518742814281371],
-	];
-
-	return multiplyMatrices(M, XYZ) as Color;
+	return multiplyMatrices(M, XYZ);
 }