+++ packages/color-helpers/src/conversions/xyz-to-lin-pro-photo.ts
@@ -1,7 +1,13 @@
 /* eslint-disable no-loss-of-precision */
 
 import type { Color } from '../types/color';
-import { multiplyMatrices } from '../calculations/multiply-matrices';
+import { Matrix, multiplyMatrices } from '../calculations/multiply-matrices';
+
+const M: Matrix = [
+	1.34578688164715830, -0.25557208737979464, -0.05110186497554526,
+	-0.54463070512490190, 1.50824774284514680, 0.02052744743642139,
+	0.00000000000000000, 0.00000000000000000, 1.21196754563894520,
+];
 
 /**
  * Convert D50 XYZ to linear-light prophoto-rgb
@@ -11,11 +17,5 @@ import { multiplyMatrices } from '../calculations/multiply-matrices';
  * @see http://www.brucelindbloom.com/index.html?Eqn_RGB_XYZ_Matrix.html
  */
 export function XYZ_to_lin_ProPhoto(XYZ: Color): Color {
-	const M = [
-		[1.34578688164715830, -0.25557208737979464, -0.05110186497554526],
-		[-0.54463070512490190, 1.50824774284514680, 0.02052744743642139],
-		[0.00000000000000000, 0.00000000000000000, 1.21196754563894520],
-	];
-
-	return multiplyMatrices(M, XYZ) as Color;
+	return multiplyMatrices(M, XYZ);
 }